Add desktop build
Need to reimage a desktop, so might as well get this in here so I
only have to do it once. For now, this is just a laptop build
without a few packages, but I expect that to change in the near
future when this machine starts doing a lot more.

Signed-off-by: Cullen Taylor <ctaylor@gitlab.com>